% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWUDS.R
\docType{package}
\name{AWUDS-package}
\alias{AWUDS-package}
\title{AWUDS}
\description{
\tabular{ll}{
Package: \tab AWUDS\cr
Type: \tab Package\cr
License: \tab Unlimited for this package, dependencies have more restrictive licensing.\cr
Copyright: \tab This software is in the public domain because it contains materials
that originally came from the United States Geological Survey, an agency of
the United States Department of Interior. For more information, see the
official USGS copyright policy at
http://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
LazyLoad: \tab yes\cr
}
}
\details{
Collection of functions to help parse and analyze AWUDS data.
}

