% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_wu.R
\name{boxplot_wu}
\alias{boxplot_wu}
\title{boxplot_wu}
\usage{
boxplot_wu(w.use, data.elements, area.column, sub.area = "COUNTYNAME",
  areas = NA, plot.notch = FALSE, years = NA, log = FALSE)
}
\arguments{
\item{w.use}{is a subset of the datafile wUseSample that includes all areas in all data elements for state}

\item{data.elements}{character name of data element within available categories by year for state}

\item{area.column}{character that defines which column to use to specify area}

\item{sub.area}{is the area (county, huc) column name. This is most useful for the app}

\item{areas}{is a geographical area as defined in your datafile such as county, HUC, or aquifer}

\item{plot.notch}{is a logical function to show boxplot as optionally notched or not}

\item{years}{vector of integers specifying range of years to graph. Defaults to NA which shows all years in dataset.}

\item{log}{= TRUE or FALSE allows user to set log scale, default is FALSE}
}
\description{
boxplot_wu
}
\examples{
w.use <- wUseSample
areas <- "15"
area.column = "STATECODE"
data.elements <- c("PS.GWPop","TP.TotPop")
boxplot_wu(w.use, data.elements, area.column = area.column,areas = areas)
boxplot_wu(w.use, data.elements, area.column, log=TRUE)
boxplot_wu(w.use, data.elements, area.column, years = c(1995, 2005))
}
