% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_values.R
\name{calculate_values}
\alias{calculate_values}
\title{caluculate_values}
\usage{
calculate_values(w.use)
}
\arguments{
\item{w.use}{dataframe, the water use data}
}
\description{
Calculate values
}
\examples{
w.use <- wUseSample
w.use.new <- calculate_values(w.use)
}
