% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_values.R
\name{caluculate_values}
\alias{caluculate_values}
\title{caluculate_values}
\usage{
caluculate_values(w.use)
}
\arguments{
\item{w.use}{dataframe, the water use data}
}
\description{
Calculate values
}
\examples{
w.use <- wUseSample
w.use.new <- caluculate_values(w.use)
}

