% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankData.R
\name{compare.data.elements}
\alias{compare.data.elements}
\title{compare.data.elements}
\usage{
compare.data.elements(file_path, file.1, file.2, data.element)
}
\arguments{
\item{file_path}{chr, path to file.1 and file.2}

\item{file.1}{character name of first file}

\item{file.2}{character name of second file}

\item{data.element}{character name of data element}
}
\description{
Compare multiple files
}
\examples{
folderPath <- system.file("extdata", package="wateRuse")
file.1 <- "Export_2010_County.xlsx"
file.2 <- "Export_2015_County.xlsx"
data.element <- "TP"
df <- compare.data.elements(folderPath, file.1, file.2, data.element)
}

