% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankData.R
\name{compare_data_elements}
\alias{compare_data_elements}
\title{compare_data_elements}
\usage{
compare_data_elements(w.use)
}
\arguments{
\item{w.use}{dataframe, the water use data}
}
\description{
Compare multiple files
}
\examples{
w.use <- wUseSample
df <- compare_data_elements(w.use)
}

