% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{explore_wateRuse}
\alias{explore_wateRuse}
\title{Run water use application}
\usage{
explore_wateRuse(browse = TRUE)
}
\arguments{
\item{browse}{use browser for map rendering}
}
\description{
Run water use application
}
\examples{
\dontrun{
explore_wateRuse()
}
}

