% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAwudsData.R
\name{get_awuds_data}
\alias{get_awuds_data}
\title{get_awuds_data}
\usage{
get_awuds_data(awuds.data.path = NA, awuds.data.files = NA)
}
\arguments{
\item{awuds.data.path}{character, path to files to try to load.}

\item{awuds.data.files}{character, a vector of files to try to load.}
}
\value{
\code{data.frame} containing AWUDS data elements
}
\description{
Get data from local excel export or dump file.
}
\examples{
awuds.data.path <- system.file("extdata/dump", package="wateRuse")
awuds.data <- get_awuds_data(awuds.data.path)

awuds.data.path <- system.file("extdata/excel_test", package="wateRuse")
awuds.data <- get_awuds_data(awuds.data.path)

fileList <- c(system.file("extdata/excel_test/Export_2005_County.xlsx", package="wateRuse"),
system.file("extdata/excel_test_test/Export_2010_County.xlsx", package="wateRuse"),
system.file("extdata/excel_test/Export_2015_County.xlsx", package="wateRuse"))
awuds.data <- get_awuds_data(awuds.data.files=fileList)

file <- c(system.file("extdata/dump/exampleAWUDSdump.txt", package="wateRuse"))
awuds.data <- get_awuds_data(awuds.data.files=file)

}
