% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wateRuse.R
\name{histCounties}
\alias{histCounties}
\title{County Polygons for use in mapping water use data.}
\description{
County polygons from: Siczewicz, Peter. U.S. Historical Counties (Generalized .001 deg). Dataset. 
Emily Kelley, digital comp. Atlas of Historical County Boundaries, ed. by John H. Long. 
Chicago: The Newberry Library, 2011. Available online from http://publications.newberry.org/ahcbp
}
\examples{
names(histCounties@data)
\dontrun{
plot(histCounties[which(histCounties@data$STATE_TERR == "New Jersey"),])
}
}
\keyword{county}
\keyword{polygons}
