% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wateRuse.R
\name{huc08Poly}
\alias{huc08Poly}
\title{HUC08 Polygons for use in mapping water use data.}
\description{
HUC08 snapshot was taken from the June 30th 2016 snapshot of the Watershed Boundary Dataset.
}
\examples{
names(huc08Poly@data)
\dontrun{
plot(huc08Poly[which(huc08Poly@data$STATES == "IA"),])
}
}
\keyword{HUC08}
\keyword{Waterhsed}
\keyword{polygons}

