% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wateRuse.R
\docType{data}
\name{lookupTables}
\alias{calculation}
\alias{calculation.formula}
\alias{category}
\alias{county}
\alias{dataelement}
\alias{huc8}
\alias{lookupTables}
\title{Included data}
\description{
\itemize{
 \item{category}{category data.frame}
 \item{county}{county data.frame}
 \item{dataelement}{dataelement data.frame}
 \item{huc8}{huc8 data.frame}
 \item{calculation}{calculation data.frame}
 \item{calculation.formula}{calculation.formula data.frame}
}
}
\examples{
head(category)
head(county)
head(dataelement)
head(huc8)
head(calculation)
head(calculation.formula)
}

