% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiElement.R
\name{multi_element_data}
\alias{multi_element_data}
\title{multi_element_data}
\usage{
multi_element_data(w.use, data.elements, area.column, plot.points = TRUE,
  years = NA, areas = NA, y.scale = NA, log = FALSE, legend = TRUE,
  c.palette = c("dodgerblue", "forestgreen"))
}
\arguments{
\item{w.use}{is a subset of the datafile wUseSample that includes all areas in all data elements for state}

\item{data.elements}{character name of data element within available categories by year for state}

\item{area.column}{character that defines which column to use to specify area}

\item{plot.points}{is a logical function to show counties as points or clustered bar graph}

\item{years}{vector of integers specifying all years available for state. Defaults to NA which shows all years in dataset.}

\item{areas}{is a geographical area as defined in your datafile such as county, HUC, or aquifer}

\item{y.scale}{allows R to set the y-axis scale given available data range. Defaults to NA which lets R set the scale based on dataset values.}

\item{log}{= TRUE or FALSE allows user to set log scale, default is FALSE}

\item{legend}{is a logical function to include list of counties in a legend if manageable, default is TRUE}

\item{c.palette}{color palette to use for points and lines}
}
\description{
multi_element_data
}
\examples{
w.use <- wUseSample
areas <- c("Kent County","Sussex County")
area.column = "COUNTYNAME"
data.elements <- c("PS.GWPop","TP.TotPop")
multi_element_data(w.use, data.elements, area.column = area.column,areas = areas)
multi_element_data(w.use, data.elements, plot.points = FALSE,
       area.column = area.column,areas = areas)
multi_element_data(w.use, data.elements, plot.points = FALSE,
       area.column = area.column,areas = "Sussex County")
multi_element_data(w.use, data.elements, area.column)
multi_element_data(w.use, data.elements, area.column, y.scale = c(0,1000))
multi_element_data(w.use, data.elements, area.column, 
       y.scale = c(0,100), years = c(1990,2005))
}

