% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDataFunctions.R
\name{parseExport}
\alias{parseExport}
\title{data parse functions}
\usage{
parseExport(file_path, citations = FALSE)
}
\arguments{
\item{file_path}{chr, path to the excel file (including file extension)}

\item{citations}{logical, citations were included as part of the output for Export data}
}
\description{
Parses AWUDS excel files into R data frames.
}

