% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_county_polygons.R
\name{subset_county_polygons}
\alias{subset_county_polygons}
\title{subset_county_polygons}
\usage{
subset_county_polygons(areas = NA, area.column, year)
}
\arguments{
\item{areas}{chr, vector of codes indicating HUCs, counties, states, regions, aquifers, etc.}

\item{area.column}{chr, defines which column to use to specify area}

\item{year}{int, the year of interest to be mapped (defines historical basis for counties)}
}
\description{
Returns county polygons from: Siczewicz, Peter. U.S. Historical Counties 
(Generalized .001 deg) on the basis of year of interest
Dataset is subset on basis of area of interest
}
\examples{
areas <- "Delaware" # should yield 3 counties
areas <- "Maine"# should yield 16 counties
areas <- "New Hampshire"# should yield 10 counties
area.column <- "STATE_TERR"
year <- 2010
subset_county_polygons(areas, area.column, year)
}

