% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_huc_polygons.R
\name{subset_huc_polygons}
\alias{subset_huc_polygons}
\title{subset_huc_polygons}
\usage{
subset_huc_polygons(year, area.column, areas)
}
\arguments{
\item{year}{int, the year of interest to be mapped (defines historical basis for HUCs)}

\item{area.column}{chr, defines which column the areas refer to for subsetting}

\item{areas}{chr, vector of codes indicating area of interest to subset on}
}
\value{
hc.sub dataframe, a subset of HUC polygon data
}
\description{
Returns HUC polygons 
Dataset is subset on basis of area of interest
}
\examples{
areas <- "ME" # 21 HUC8s in Maine 
area.column <- "STATES"
year <- 1995 
hc.sub <- subset_huc_polygons(year, area.column, areas)
hc.sub$HUCNAME
}
