% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_wuse.R
\name{subset_wuse}
\alias{subset_wuse}
\title{subset_wuse}
\usage{
subset_wuse(w.use, data.elements, area.column, areas = NA)
}
\arguments{
\item{w.use}{dataframe, the water use data}

\item{data.elements}{chr, vector of data elements to be plotted (1 plot per element)}

\item{area.column}{character that defines which column to use to specify area}

\item{areas}{chr, codes for state-county ('NA' uses all areas)}
}
\value{
w.use.sub dataframe, a subset of water use data from user specifications
}
\description{
returns subset of WUSE data elements for plotting based on county state codes
}
\examples{
w.use <- wUseSample
data.elements <- c("PS.TOPop", "PS.SWPop")
areas <- "10" # NA uses all areas
area.column <- "STATECODE"

w.use.sub <- subset_wuse(w.use, data.elements,area.column,areas)
}
