% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{time_series_data}
\alias{time_series_data}
\title{time.series.data}
\usage{
time_series_data(data.element, years = NA)
}
\arguments{
\item{data.element}{character name of data element within available categories by year for state}

\item{years}{vector of integers specifying all years available for state}
}
\description{
time.series.data
}
\examples{

data.element <- "PS-GWPop"
time_series_data(data.element)
}

