% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wateRuse.R
\docType{package}
\name{wateRuse-package}
\alias{wateRuse-package}
\title{wateRuse}
\description{
\tabular{ll}{
Package: \tab wateRuse\cr
Type: \tab Package\cr
License: \tab Unlimited for this package, dependencies have more restrictive licensing.\cr
Copyright: \tab This software is in the public domain because it contains materials
that originally came from the United States Geological Survey, an agency of
the United States Department of Interior. For more information, see the
official USGS copyright policy at
https://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
LazyLoad: \tab yes\cr
}
}
\details{
Collection of functions to help parse and analyze AWUDS data.
}
