% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wateRuse.R
\docType{data}
\name{wUseSample}
\alias{wUseSample}
\title{Example Water Use Data}
\description{
Example dump file read in as RData format.
}
\examples{
head(wUseSample)
}
\keyword{data}
\keyword{use}
\keyword{water}

