% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyPlot.R
\name{xyPlot}
\alias{xyPlot}
\title{xyPlot}
\usage{
xyPlot(w.use, x_arg, y_arg, z_subset)
}
\arguments{
\item{x_arg}{chr, the x-axis parameter to be plotted}

\item{y_arg}{chr, the y-axis parameter to be plotted}

\item{z_subset}{chr, subsetting criteria}

\item{df}{data}
}
\description{
returns scatter plot for user-defined params and subset criteria
}
\examples{
x_arg <- "IR_Totwdl"
y_arg <- "IR_Totacr"
z_subset <- c("county code", "state code", "2000")
plot <- xyPlot(x_arg,y_arg,z_subset)
plot
}

