% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_plot.R
\name{choropleth_plot}
\alias{choropleth_plot}
\title{choropleth_plot}
\usage{
choropleth_plot(data.elements, norm.element = NA, year, areas, area.column)
}
\arguments{
\item{data.elements}{chr, data element to be plotted}

\item{norm.element}{chr, data element to be used for normalizing data.elements}

\item{year}{int, the year of interest to be mapped (defines historical basis for counties)}

\item{areas}{chr, vector of codes indicating HUCs, counties, states, regions, aquifers, etc.}

\item{area.column}{chr, defines which column to use to specify area}
}
\description{
Returns choropleth plot of data element of interest for year of interest
Currently written to work with counties only
Only works with one state, data element, and year as currently written
}
\examples{
w.use <- wUseSample
data.elements <- "PS.WFrTo"
norm.element <- "PS.TOPop"
year <- 2010 
areas <- "Delaware" 
area.column <- "STATE_TERR"
ch.plot <- choropleth_plot(data.elements, norm.element, year, areas, area.column)
ch.plot
norm.element <- NA
ch.plot <- choropleth_plot(data.elements, norm.element, year, areas, area.column)
ch.plot
}

