% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_plot.R
\name{choropleth_plot}
\alias{choropleth_plot}
\title{choropleth_plot}
\usage{
choropleth_plot(w.use, data.elements, year, state, norm.element = NA,
  unit.type = "county")
}
\arguments{
\item{w.use}{dataframe, the water use data}

\item{data.elements}{chr, data element to be plotted}

\item{year}{int, the year of interest to be mapped (defines historical basis for counties)}

\item{state}{character name of state}

\item{norm.element}{chr, data element to be used for normalizing data.elements}

\item{unit.type}{chr, type of unit to be mapped; acceptable options are "county", "huc", or "aquifer"}
}
\description{
Returns choropleth plot of data element of interest for year of interest
Currently written to work with counties and hucs only
Only works with one data element and year as currently written
}
\examples{
w.use <- wUseSample
data.elements <- "PS.WFrTo"
norm.element <- "PS.TOPop"
year <- 2010 
state <- "Delaware" 
ch.plot <- choropleth_plot(w.use, data.elements, year, state, norm.element)
ch.plot
norm.element <- NA
ch.plot <- choropleth_plot(w.use, data.elements, year, state, norm.element)
ch.plot

}
