% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_two_years.R
\name{compare_two_years}
\alias{compare_two_years}
\title{compare_two_years}
\usage{
compare_two_years(w.use, data.elements, year.x.y, area.column,
  areas = NA, legend = FALSE, pctDiff = 0, valDiff = 0,
  log.oper = "and", c.palette = c("#999999", "#E69F00", "#56B4E9",
  "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7"))
}
\arguments{
\item{w.use}{dataframe, the water use data}

\item{data.elements}{chr, vector of data elements to be plotted (1 plot per element)}

\item{year.x.y}{int, 2-element vector specifying the two years to be plotted}

\item{area.column}{character that defines which column to use to specify area}

\item{areas}{chr, codes indicating HUCs, counties, states, aquifers, etc.}

\item{legend}{is a logical function to include list of counties in a legend if manageable, default is FALSE}

\item{pctDiff}{is the percent difference outlier criterion , default is zero}

\item{valDiff}{is the value difference outlier criterion, default is zero}

\item{log.oper}{indictaes the logical operator between the two outlier criteria, default is "and"}

\item{c.palette}{color palette to use for points}
}
\description{
Returns scatter plot for 2 years for data elements of interest.
Option to narrow down data geographically
}
\examples{
w.use <- wUseSample
data.elements <- c("PS.TOPop", "PS.SWPop")
areas <- "10" # NA uses all areas
area.column <- "STATECODE"
year.x.y <- c(2005,2010)
compare_two_years(w.use, data.elements, year.x.y, area.column, areas)
compare_two_years(w.use, data.elements, year.x.y, area.column)
compare_two_years(w.use, "PS.TOPop", year.x.y, area.column)
pctDiff <- 50
valDiff <- 100
compare_two_years(w.use, "PS.TOPop", year.x.y, area.column, pctDiff, valDiff)
}
