% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAwudsData.R
\name{get_awuds_data}
\alias{get_awuds_data}
\title{get_awuds_data}
\usage{
get_awuds_data(awuds.data.path = NA)
}
\arguments{
\item{awuds.data.path}{character, path to files}
}
\value{
\code{data.frame} containing AWUDS data elements
}
\description{
Get data from local excel export or dump file.
}
\examples{
awuds.data.path <- system.file("extdata", package="wateRuse")
awuds.data <- get_awuds_data(awuds.data.path)
}

