% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDataFunctions.R
\name{parseExport}
\alias{parseCompareData}
\alias{parseEnteredElements}
\alias{parseExport}
\title{data parse functions}
\usage{
parseExport(file_path, citations = FALSE)

parseEnteredElements(file_path)

parseCompareData(file_path)
}
\arguments{
\item{file_path}{chr, path to the excel file (including file extension)}

\item{citations}{logical, citations were included as part of the output for Export data}
}
\description{
Parses AWUDS excel files into R data frames.
}
\examples{
path <- system.file("extdata", package="AWUDS")
exportData <- parseExport(file.path(path,"Export_2010_County.xlsx"),citation=TRUE)
TP <- exportData[["TP"]]
PO <- exportData[["PO"]]
path <- system.file("extdata", package="AWUDS")
path <- system.file("extdata", package="AWUDS")
compareData <- parseCompareData(file.path(path, "CompareData.xlsx"))
}

