% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{time_series_data}
\alias{time_series_data}
\title{time_series_data}
\usage{
time_series_data(w.use, data.elements, years = NA, area.column = NA,
  areas = NA, y.scale = NA, log = FALSE)
}
\arguments{
\item{w.use}{is the datafile that includes all areas in all data elements for state}

\item{data.elements}{character name of data element within available categories by year for state}

\item{years}{vector of integers specifying all years available for state. Defaults to NA which shows all years in dataset.}

\item{area.column}{character that defines which column to use to specify area}

\item{areas}{is a geographical area as defined in your datafile such as county, HUC, or aquifer}

\item{y.scale}{allows R to set the y-axis scale given available data range. Defaults to NA which lets R set the scale based on dataset values.}

\item{log}{= TRUE or FALSE allows user to set log scale, default is FALSE}
}
\description{
time_series_data
}
\examples{
w.use <- wUseSample
year1 <- 2005
year2 <- 2010
years <- c(year1, year2)
areas <- c("Kent County","Sussex County")
data.elements <- "PS.GWPop"
plotObject <- time_series_data(w.use, data.elements, area.column = "COUNTYNAME",areas = areas)
plotObject <- time_series_data(w.use, data.elements)
plotObject <- time_series_data(w.use, data.elements, y.scale = c(0,100))
plotObject <- time_series_data(w.use, data.elements, y.scale = c(0,100), years = c(1990,2005))
}

