% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wateRuse.R
\docType{data}
\name{wUseSample}
\alias{wUseSample}
\title{Example Water Use Data}
\description{
Example dump file read in as RData format.
}
\details{
This file is modified from a dataset retrieved from the U.S. Geological Survey Aggregated Water Use Database (AWUDS) on April 27, 2016, and may have been revised since this date. The data are approved and contains county-level data for Hawaii and Delaware. The file contains data for multiple years, and all USGS categories of water use data.
}
\examples{
head(wUseSample)
}
\keyword{data}
\keyword{use}
\keyword{water}

